requirements.txt
fastapi==0.78.0
fastapi_camelcase==1.0.5
itsdangerous==2.1.2
lxml==4.8.0
passlib==1.7.4
pdf417decoder==1.0.8
Pillow==9.1.1
py_linq==1.3.0
pydantic==1.8.2
PyJWT==2.4.0
python-dotenv==0.20.0
python_dateutil==2.8.2
requests==2.26.0
SQLAlchemy==1.4.35
sqlalchemy_mixins==1.5
starlette
structlog==21.5.0
uvicorn==0.14.0
